feat(device): USB I/O control slice 3 — Windows-first (issue #137) (#147)
* feat(device): USB I/O control slice 3 (Windows-first, issue #137)

Add opt-in USB enumerate/open/read/write/close behind device.usb.io cap,
trust gate, and confirmation mirroring device capture. Windows backend uses
SetupAPI enumeration and COM serial I/O; Linux/macOS return
UnsupportedPlatform. Includes FakeUsbIoBackend CI tests, agent tools, grant
chrome hooks (USB device label), and docs.


* fix(ui): shorten USB grant dialog title to « USB »

Mirror Camera/Micro brevity in confirm_device_usb chrome: EN/FR title
is now « USB » instead of « USB device » / « Périphérique USB ».
Button labels unchanged. Docs updated in device-usb.md and TESTER.


* fix(packaging): refresh mcp-example catalogue hash after main rebase

Rebase onto main picked up 61c2a6a changes to share/mcp/servers.yaml.example
(headless-browser MCP docs) without an updated catalogue entry. Set mcp-example
sha256 to 29c7b173… and regenerate catalogue.yaml.sig via UPDATE_CATALOGUE=1.


* chore(release): bump version to 0.16.2 and update documentation

- Updated version numbers across Cargo.lock, Cargo.toml, and VERSION files to 0.16.2.
- Revised README, INSTALL, and other documentation to reflect the new version and features.
- Introduced USB I/O functionality for Windows, allowing serial communication with USB devices.
- Updated community and feature documentation to highlight changes in the new preview.


* fix(chat): auto-inject device.usb tools on USB list intent

Mirror camera/mic capture delegation for USB device queries:
- Detect FR/EN USB/serial intents in chat_delegate (with long-doc guard)
- push_device_usb_tools wires enumerate/open/read/write/close into delegate kits
- Canonicalize common usb.* hallucinations to device.usb.*
- Reserve usb/shell prefixes so they never hit WASM module fallback


* fix(ui): satisfy clippy after main merge (files UI, billing)

- Derive Default for FilesUiState; use struct-update in tests
- Replace format!("{path}") with path.to_string() in files ops
- Drop redundant now_ms() as u64 casts
- Allow dead_code on billing ProviderRate::known (phase-1 field)


* refactor(ui): streamline agent tool selection with helper function

Replaced manual checkbox handling for agent tools with a dedicated `ui_roster_tool_checkboxes` function, improving code readability and maintainability. This change simplifies the UI logic for managing tool selections in the agent interface.

* fix(ui): IPv6 loopback detection and backup restore path base

Parse bracketed IPv6 hosts in endpoint_is_loopback so [::1] URLs are free.

Use the destination directory as rel_base during restore copy_dir_recursive;
the backup source path caused strip_prefix to fail with chemin hors backup.


* fix(usb): reject device.usb.io cap name and broaden COM intent

Map device.usb.io with open args to device.usb.open; otherwise return a clear
error naming enumerate/open. Detect COM port and se connecter phrases for USB
tool injection. Add regression tests.


* fix(agent): preserve goal.complete result when summary is empty

Require a human-readable summary in goal.complete tool schema and prompts.

Add resolve_goal_complete_summary to fall back to thought or the prior tool
result instead of only terminé. aos-agentd no longer clears last_output on
goal.complete steps with an empty summary.


* fix(usb): brief directive for COM connect and export completion body

When chat intent is connect/open/COM, append a mandatory enumerate→open
directive to the delegate brief after injecting USB tools. List-only USB
queries skip the directive.

Strengthen device.usb.io rejection (shell.run, device.enumerate called out).
Session export labels agent completion attachments and includes message body.


* fix(usb): show device.usb.open in grant banner, not cap id

Pass device.usb.open to policy_gate for pending confirmations while
keeping action.kind=device.usb.io for policy matching. Map device.usb.io
to localized roster labels in the confirm UI as a fallback.


* fix(usb): human FR hint when COM port is busy

Do not treat périphérique occupé as a missing-cap PermissionDenied.
Suggest closing the other app instead of cap.request tool.invoke:device.
Use device.usb.io for genuine USB cap-request hints.


* fix(agent): prefer tool output over internal goal.complete thoughts

When summary is empty and thought is meta (e.g. Aucune action
supplémentaire), fall back to the prior user-visible tool result.


---------

+++ crates/aos-ui-egui/src/ui_models.rs
@@ -267,7 +267,7 @@ impl UiApp {
                                             models_disk::note_used(
                                                 &mut self.models_ui.model_usage,
                                                 &id,
-                                                crate::now_ms() as u64,
+                                                crate::now_ms(),
                                             );
                                             let _ = self.cmd_tx.send(Cmd::ModelLoad {
                                                 model_id: id.clone(),
@@ -284,7 +284,7 @@ impl UiApp {
                                             models_disk::note_used(
                                                 &mut self.models_ui.model_usage,
                                                 &id,
-                                                crate::now_ms() as u64,
+                                                crate::now_ms(),
                                             );
                                             let _ = self.cmd_tx.send(Cmd::ModelReload {
                                                 model_id: id.clone(),
